[Silver V] Title: 불사조, Time: 176 ms, Memory: 45196 KB -BaekjoonHub

+++ 백준/Silver/31780. 불사조/불사조.java
@@ -17,7 +17,7 @@ public class Main {
         Queue<Integer> q = new ArrayDeque<>();
         q.offer(x);
         
-        long sum = x;
+        int sum = x;
         for(int i = 0; i < m; i++) {
         	int size = q.size();
